Feat: 검색결과 페이지
- 스페이스는 공통으로 사용해서 component로 뺐음
