+++ src/App.js
@@ -2,7 +2,9 @@
 
 import axios from "axios";
 import React, { useState, useEffect } from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Dashboard from "./pages/dashBoard";
+import Search from "./pages/search";
 
 function App() {
   const sendRequest = async () => {
@@ -16,9 +18,13 @@ function App() {
   });
 
   return (
-    <div className="App">
-      <Dashboard />;
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/dash-board" element={<Dashboard />} />{" "}
+        {/* 대시보드 페이지 */}
+        <Route path="/search" element={<Search />} /> {/* 검색 페이지 */}
+      </Routes>
+    </Router>
   );
 }
